@@ -1,14 +1,26 @@
+from __future__ import annotations
+
+import typing
+from collections.abc import Sequence
 from copy import copy
 from enum import Enum, IntEnum
 from pathlib import Path
-from typing import Literal, Optional, Self, Tuple, overload
+from types import EllipsisType
+from typing import TYPE_CHECKING, Literal, Optional, Self, Tuple, TypeAlias, overload
 
 import numpy as np
 import numpy.typing as npt
 
 from .utils.geometric import Point
-from .utils.image import crop_pad_center, find_centroid, read_image, resize
+from .utils.image import crop_pad_center, find_centroid, label_map_to_rgb, read_image, resize, write_image
 from .utils.safe_import import is_torch_tensor
+
+if TYPE_CHECKING:
+    import torch
+
+    from .utils.typing import PathLike
+
+    type ImageSource = torch.Tensor | npt.NDArray | PathLike | Sequence[PathLike]
 
 ABSENT = Point(float("nan"), float("nan"))
 
@@ -26,7 +38,7 @@ class ReshapeMethod(str, Enum):
     RAISE = "raise"
 
     @classmethod
-    def parse(cls, value: str) -> Self:
+    def parse(cls, value: ReshapeMethods) -> Self:
         """Parse a string to a ReshapeMethod enum."""
         if isinstance(value, cls):
             return value
@@ -36,6 +48,9 @@ class ReshapeMethod(str, Enum):
             raise ValueError(
                 f"Invalid reshape method: {value}.\nValid options are: {', '.join(m.value for m in ReshapeMethod)}"
             ) from e
+
+
+ReshapeMethods: TypeAlias = Literal["resize", "crop", "raise"] | ReshapeMethod
 
 
 class AVLabel(IntEnum):
@@ -91,9 +106,9 @@ class AVLabel(IntEnum):
 class FundusData:
     def __init__(
         self,
-        fundus=None,
+        image=None,
         *,
-        fundus_mask=None,
+        roi_mask=None,
         vessels=None,
         av=None,
         od=None,
@@ -104,35 +119,36 @@ class FundusData:
         name: Optional[str] = None,
         check_validity: bool = True,
         target_shape: Optional[Tuple[int, int]] = None,
-        reshape_method: ReshapeMethod = ReshapeMethod.RAISE,
+        reshape_method: ReshapeMethods = ReshapeMethod.RAISE,
+        immutable: bool = False,
     ):
         shape: Optional[Tuple[int, int]] = target_shape
-        if fundus is not None:
+        if image is not None:
             if check_validity:
-                fundus = self.load_fundus_image(fundus, target_shape, reshape_method=reshape_method)
-            self._fundus = fundus
+                image = self.load_fundus_image(image, target_shape, reshape_method=reshape_method)
+            self._image = image
             if shape is None:
-                shape = self._fundus.shape[:2]  # type: ignore[assignment]
+                shape = self._image.shape[-2:]  # type: ignore[assignment]
         else:
-            self._fundus = None
+            self._image = None
 
-        if fundus_mask is not None:
+        if roi_mask is not None:
             if check_validity:
-                fundus_mask = self.load_fundus_mask(
-                    fundus_mask, from_fundus=False, target_shape=shape, reshape_method=reshape_method
+                roi_mask = self.load_fundus_mask(
+                    roi_mask, from_fundus=False, target_shape=shape, reshape_method=reshape_method
                 )
-            self._fundus_mask = fundus_mask
+            self._fundus_mask = roi_mask
             if shape is None:
-                shape = fundus_mask.shape[:2]  # type: ignore[assignment]
-        elif self._fundus is not None:
-            self._fundus_mask = self.load_fundus_mask(self._fundus, from_fundus=True)
+                shape = roi_mask.shape[-2:]  # type: ignore[assignment]
+        elif self._image is not None:
+            self._fundus_mask = self.load_fundus_mask(self._image, from_fundus=True)
 
         if vessels is not None:
             if check_validity:
                 vessels = self.load_vessels(vessels, target_shape=shape, reshape_method=reshape_method)
             self._vessels = vessels
             if shape is None:
-                shape = vessels.shape[:2]  # type: ignore[assignment]
+                shape = vessels.shape[-2:]  # type: ignore[assignment]
         else:
             self._vessels = None
         self._bin_vessels = None
@@ -142,7 +158,7 @@ class FundusData:
                 av = self.load_av(av, target_shape=shape, reshape_method=reshape_method)
             self._av = av
             if shape is None:
-                shape = av.shape[:2]  # type: ignore[assignment]
+                shape = av.shape[-2:]  # type: ignore[assignment]
         else:
             self._av = None
 
@@ -152,7 +168,7 @@ class FundusData:
                     od, shape, reshape_method=reshape_method, fit_ellipse=True
                 )  # type: ignore
                 if shape is None:
-                    shape = self._od.shape[:2]  # type: ignore[assignment]
+                    shape = self._od.shape[-2:]  # type: ignore[assignment]
             else:
                 self._od, self._od_center, self._od_size = od, od_center, od_size
         else:
@@ -162,7 +178,7 @@ class FundusData:
             if check_validity:
                 self._macula, self._macula_center = self.load_od_macula(macula, shape, reshape_method=reshape_method)
                 if shape is None:
-                    shape = self._macula.shape[:2]  # type: ignore[assignment]
+                    shape = self._macula.shape[-2:]  # type: ignore[assignment]
             else:
                 self._macula, self._macula_center = macula, macula_center
         else:
@@ -173,8 +189,8 @@ class FundusData:
         self._shape = shape
 
         if name is None:
-            if isinstance(fundus, (str, Path)):
-                name = Path(fundus).stem
+            if isinstance(image, (str, Path)):
+                name = Path(image).stem
             elif isinstance(vessels, (str, Path)):
                 name = Path(vessels).stem
             elif isinstance(od, (str, Path)):
@@ -182,39 +198,80 @@ class FundusData:
             elif isinstance(macula, (str, Path)):
                 name = Path(macula).stem
         self._name = name
+        self._immutable = immutable
+
+    type Fields = Literal["image", "fundus_mask", "vessels", "av", "od", "macula"]
 
     def update(
         self,
-        fundus=None,
-        fundus_mask=None,
-        vessels=None,
-        av=None,
-        od=None,
-        macula=None,
-        reshape_method: ReshapeMethod = ReshapeMethod.RAISE,
-        name: Optional[str] = None,
+        image: ImageSource | EllipsisType = ...,
+        fundus_mask: ImageSource | EllipsisType = ...,
+        vessels: ImageSource | EllipsisType = ...,
+        av: ImageSource | EllipsisType = ...,
+        od: ImageSource | EllipsisType = ...,
+        macula: ImageSource | EllipsisType = ...,
+        target_shape: Optional[Tuple[int, int]] = None,
+        reshape_method: ReshapeMethods = ReshapeMethod.RAISE,
+        name: str | EllipsisType = ...,
+        inplace: bool = False,
     ) -> Self:
-        other = copy(self)
-        shape = self.shape
-        if fundus is not None:
-            other._fundus = other.load_fundus_image(fundus, target_shape=shape, reshape_method=reshape_method)
-        if fundus_mask is not None:
+        if inplace:
+            if self._immutable:
+                raise RuntimeError("This FundusData instance is immutable.")
+            other = self
+        else:
+            other = copy(self)
+
+        shape = self.shape if target_shape is None else target_shape
+        if image is not ...:
+            other._image = other.load_fundus_image(image, target_shape=shape, reshape_method=reshape_method)
+        if fundus_mask is not ...:
             other._fundus_mask = other.load_fundus_mask(fundus_mask, target_shape=shape, reshape_method=reshape_method)
-        if vessels is not None:
+        if vessels is not ...:
             other._vessels = other.load_vessels(vessels, target_shape=shape, reshape_method=reshape_method)
-        if av is not None:
+        if av is not ...:
             other._av = other.load_av(av, target_shape=shape, reshape_method=reshape_method)
 
-        if od is not None:
+        if od is not ...:
             other._od, other._od_center, other._od_size = other.load_od_macula(
                 od, target_shape=shape, reshape_method=reshape_method, fit_ellipse=True
             )
-        if macula is not None:
+        if macula is not ...:
             other._macula, other._macula_center = other.load_od_macula(
                 macula, target_shape=shape, reshape_method=reshape_method
             )
-        if name is not None:
+        if name is not ...:
             other._name = name
+        return other
+
+    def to_immutable(self) -> Self:
+        """Return an immutable copy of this FundusData."""
+        if self._immutable:
+            return self
+        other = copy(self)
+        other._set_immutable_flag(True)
+        return other
+
+    def _set_immutable_flag(self, immutable: bool) -> None:
+        """Set the immutable flag to True."""
+        self._immutable = immutable
+        if self._image is not None:
+            self._image.setflags(write=not immutable)
+        if self._fundus_mask is not None:
+            self._fundus_mask.setflags(write=not immutable)
+        if self._vessels is not None:
+            self._vessels.setflags(write=not immutable)
+        if self._av is not None:
+            self._av.setflags(write=not immutable)
+        if self._od is not None:
+            self._od.setflags(write=not immutable)
+        if self._macula is not None:
+            self._macula.setflags(write=not immutable)
+
+    def mutable_copy(self) -> Self:
+        """Return a mutable copy of this FundusData."""
+        other = copy(self)
+        other._set_immutable_flag(False)
         return other
 
     def remove_od_from_vessels(self):
@@ -235,123 +292,163 @@ class FundusData:
     @classmethod
     def load_fundus_image(
         cls,
-        fundus,
+        fundus: ImageSource,
         target_shape: Optional[Tuple[int, int]] = None,
-        reshape_method: ReshapeMethod = ReshapeMethod.RAISE,
+        reshape_method: ReshapeMethods = ReshapeMethod.RAISE,
     ) -> npt.NDArray[np.float32]:
+        """Load a fundus image from various sources.
+        The image is converted to a numpy array of type float32, with channel first (3, H, W) and values in [0, 1].
+
+        Parameters
+        ----------
+        fundus : ImageSource
+            The fundus image. It must be one of:
+            - A path to an image file containing the fundus image.
+            - A numpy array (or torch Tensor) of shape (H, W, 3) or (3, H, W) containing the fundus image.
+        target_shape : Optional[Tuple[int, int]], optional
+            The target shape of the image, by default None
+        reshape_method : ReshapeMethods, optional
+            The method to use for reshaping the image, by default ReshapeMethod.RAISE. If target_shape is None, this parameter is ignored.
+
+        Returns
+        -------
+        npt.NDArray[np.float32]
+            The loaded fundus image as a numpy array.
+
+        Raises
+        ------
+        TypeError
+            If the fundus image is not a path, a numpy array or a torch tensor.
+        ValueError
+            If the fundus image is not a color image.
+        """  # noqa: E501
         # --- Load the image ---
         if isinstance(fundus, (str, Path)):
-            fundus = read_image(fundus)
+            fundus_ = read_image(fundus)
         elif is_torch_tensor(fundus):
-            fundus = fundus.numpy(force=True)
+            fundus_ = fundus.numpy(force=True)
+        elif isinstance(fundus, np.ndarray):
+            fundus_ = fundus
+        else:
+            raise TypeError("The fundus image must be a path, a numpy array or a torch tensor.")
 
         # --- Check image ---
-        assert isinstance(fundus, np.ndarray), "The image must be a numpy array."
-        assert fundus.ndim == 3 and fundus.shape[2] == 3, "The image must be a color image."
+        assert fundus_.ndim == 3, "The image must be a color image."
+        if fundus_.shape[0] != 3 and fundus_.shape[2] == 3:
+            fundus_ = fundus_.transpose(2, 0, 1)  # HWC -> CHW
+        assert fundus_.shape[0] == 3, "The image must be a RGB color image."
 
         # --- Format data ---
-        if fundus.dtype != np.float32:
-            fundus = fundus.astype(np.float32)
+        if fundus_.dtype != np.float32:
+            fundus_ = fundus_.astype(np.float32)
 
         # -- Resize the image ---
-        if target_shape is not None and target_shape != fundus.shape[:2]:
+        if target_shape is not None and target_shape != fundus_.shape[-2:]:
             match ReshapeMethod.parse(reshape_method):
                 case ReshapeMethod.RESIZE:
-                    fundus = resize(fundus, target_shape, interpolation=True)
+                    fundus_ = resize(fundus_, target_shape, interpolation=True)
                 case ReshapeMethod.CROP:
-                    fundus = crop_pad_center(fundus, target_shape)
+                    fundus_ = crop_pad_center(fundus_, target_shape)
                 case ReshapeMethod.RAISE:
                     raise ValueError(
-                        f"The fundus image shape {fundus.shape} differs from the target shape {target_shape}."
+                        f"The fundus_ image shape {fundus_.shape} differs from the target shape {target_shape}."
                     )
-        return fundus  # type: ignore
+        return fundus_  # type: ignore
 
     @classmethod
     def load_fundus_mask(
         cls,
-        fundus_mask,
+        fundus_mask: ImageSource,
         from_fundus: bool = False,
         target_shape: Optional[Tuple[int, int]] = None,
-        reshape_method: ReshapeMethod = ReshapeMethod.RAISE,
+        reshape_method: ReshapeMethods = ReshapeMethod.RAISE,
     ) -> npt.NDArray[np.bool_]:
         if from_fundus:
             # --- Compute from fundus mask ---
             from .utils.fundus import fundus_ROI
 
             fundus = cls.load_fundus_image(fundus_mask, target_shape, reshape_method=reshape_method)
-            fundus_mask = fundus_ROI(fundus)  # Compute the fundus mask from the fundus image
+            fundus_mask_ = fundus_ROI(fundus)  # Compute the fundus mask from the fundus image
         else:
             # --- Load the image ---
             if isinstance(fundus_mask, (str, Path)):
-                fundus_mask = read_image(fundus_mask, cast_to_float=False)
+                fundus_mask_ = read_image(fundus_mask, cast_to_float=False)
             elif is_torch_tensor(fundus_mask):
-                fundus_mask = fundus_mask.numpy(force=True)
+                fundus_mask_ = fundus_mask.numpy(force=True)
+            elif isinstance(fundus_mask, np.ndarray):
+                fundus_mask_ = fundus_mask
+            else:
+                raise TypeError("The fundus mask must be a path, a numpy array or a torch tensor.")
 
             # --- Check image ---
-            assert isinstance(fundus_mask, np.ndarray), "The image must be a numpy array."
-            if fundus_mask.ndim != 2:
+            if fundus_mask_.ndim != 2:
                 raise ValueError("Invalid fundus mask")
 
-            if fundus_mask.dtype != bool:
-                fundus_mask = fundus_mask > 127 if fundus_mask.dtype == np.uint8 else fundus_mask > 0.5
+            if fundus_mask_.dtype != bool:
+                fundus_mask_ = fundus_mask_ > 127 if fundus_mask_.dtype == np.uint8 else fundus_mask_ > 0.5
 
         # -- Resize the image ---
-        if target_shape is not None and target_shape != fundus_mask.shape[:2]:
+        if target_shape is not None and target_shape != fundus_mask_.shape[-2:]:
             match ReshapeMethod.parse(reshape_method):
                 case ReshapeMethod.RESIZE:
-                    fundus_mask = resize(fundus_mask, target_shape, interpolation=False)
+                    fundus_mask_ = resize(fundus_mask_, target_shape, interpolation=False)
                 case ReshapeMethod.CROP:
-                    fundus_mask = crop_pad_center(fundus_mask, target_shape)
+                    fundus_mask_ = crop_pad_center(fundus_mask_, target_shape)
                 case ReshapeMethod.RAISE:
                     raise ValueError(
-                        f"The fundus image shape {fundus_mask.shape} differs from the target shape {target_shape}."
+                        f"The fundus image shape {fundus_mask_.shape} differs from the target shape {target_shape}."
                     )
-        return fundus_mask  # type: ignore
+        return fundus_mask_  # type: ignore
 
     @classmethod
     def load_vessels(
         cls,
-        vessels,
+        vessels: ImageSource,
         target_shape: Optional[Tuple[int, int]] = None,
-        reshape_method: ReshapeMethod = ReshapeMethod.RAISE,
+        reshape_method: ReshapeMethods = ReshapeMethod.RAISE,
     ) -> npt.NDArray[np.bool_]:
         # --- Load the image ---
         if isinstance(vessels, (str, Path)):
-            vessels = read_image(vessels, cast_to_float=False)
+            vessels_ = read_image(vessels, cast_to_float=False)
         elif is_torch_tensor(vessels):
-            vessels = vessels.numpy(force=True)
+            vessels_ = vessels.numpy(force=True)
+        elif isinstance(vessels, np.ndarray):
+            vessels_ = vessels
+        else:
+            raise TypeError("The vessels segmentation must be a path, a numpy array or a torch tensor.")
 
         # --- Check image ---
-        assert isinstance(vessels, np.ndarray), "The image must be a numpy array."
-        if vessels.ndim == 3:
+        if vessels_.ndim == 3:
+            if vessels_.shape[0] not in (1, 3) and vessels_.shape[2] in (1, 3):
+                vessels_ = vessels_.transpose(2, 0, 1)  # HWC -> CHW
             # Check if the image is a binary image
-            MAX = 255 if vessels.dtype == np.uint8 else 1
-            vessels = np.any(vessels > MAX / 2, axis=2)
-        elif vessels.ndim != 2:
+            MAX = 255 if vessels_.dtype == np.uint8 else 1
+            vessels_ = np.any(vessels_ > MAX / 2, axis=0)
+        elif vessels_.ndim != 2:
             raise ValueError("Invalid fundus mask")
 
-        if vessels.dtype != bool:
-            vessels = vessels > 127 if vessels.dtype == np.uint8 else vessels > 0.5
+        if vessels_.dtype != bool:
+            vessels_ = vessels_ > 127 if vessels_.dtype == np.uint8 else vessels_ > 0.5
 
         # -- Resize the image ---
-        if target_shape is not None and target_shape != vessels.shape[:2]:
+        if target_shape is not None and target_shape != vessels_.shape[-2]:
             match ReshapeMethod.parse(reshape_method):
                 case ReshapeMethod.RESIZE:
-                    vessels = resize(vessels, target_shape, interpolation=True)
+                    vessels_ = resize(vessels_, target_shape, interpolation=True)
                 case ReshapeMethod.CROP:
-                    vessels = crop_pad_center(vessels, target_shape)
+                    vessels_ = crop_pad_center(vessels_, target_shape)
                 case ReshapeMethod.RAISE:
                     raise ValueError(
-                        f"The fundus image shape {vessels.shape} differs from the target shape {target_shape}."
+                        f"The fundus image shape {vessels_.shape} differs from the target shape {target_shape}."
                     )
-        return vessels  # type: ignore
+        return vessels_  # type: ignore
 
     @classmethod
     def load_av(
         cls,
-        av,
+        av: ImageSource,
         target_shape: Optional[Tuple[int, int]] = None,
-        reshape_method: ReshapeMethod = ReshapeMethod.RAISE,
+        reshape_method: ReshapeMethods = ReshapeMethod.RAISE,
         invert_av: bool = False,
     ) -> npt.NDArray[np.uint8]:
         """Load the arteries and veins map.
@@ -392,115 +489,127 @@ class FundusData:
         """  # noqa: E501
         # --- Load the image ---
         if isinstance(av, (str, Path)):
-            av = read_image(av, cast_to_float=False)
+            av_ = read_image(av, cast_to_float=False)
         elif is_torch_tensor(av):
-            av = av.numpy(force=True)
+            av_ = av.numpy(force=True)
+        elif isinstance(av, np.ndarray):
+            av_ = av
+        else:
+            raise TypeError("The vessels segmentation must be a path, a numpy array or a torch tensor.")
 
         # --- Interpret image ---
-        assert isinstance(av, np.ndarray), "The vessels map must be a numpy array."
-        if av.ndim == 3:
-            # Check if the image is a binary image
-            MAX = 255 if av.dtype == np.uint8 else 1
-            av = av > MAX / 2
+        assert isinstance(av_, np.ndarray), "The vessels map must be a numpy array."
+        if av_.ndim == 3:
+            if av_.shape[0] not in (1, 3) and av_.shape[2] in (1, 3):
+                av_ = av_.transpose(2, 0, 1)  # HWC -> CHW
 
-            if np.all(np.all(av, axis=2)):
-                av = (av[:, :, 0] * AVLabel.UNK).astype(np.uint8)
+            # Check if the image is a binary image
+            MAX = 255 if av_.dtype == np.uint8 else 1
+            av_ = av_ > MAX / 2
+
+            if np.all(np.all(av_, axis=2)):
+                av_ = (av_[0] * AVLabel.UNK).astype(np.uint8)
             else:
-                av_map = np.zeros(av.shape[:2], dtype=np.uint8)
+                av_map = np.zeros(av_.shape[-2:], dtype=np.uint8)
                 if invert_av:
-                    av = av[:, :, ::-1]  # Invert the channels: red (artery) and blue (vein)
-                av_map[av[:, :, 0]] = AVLabel.ART
-                av_map[av[:, :, 2]] = AVLabel.VEI
-                av_map[av[:, :, 1]] = AVLabel.BOTH
-                av_map[av[:, :, 2] & av[:, :, 0]] = AVLabel.BOTH
-                av = av_map
-        elif av.ndim == 2:
-            if av.dtype == bool:
-                av = (av * AVLabel.UNK).astype(np.uint8)
-            elif np.issubdtype(av.dtype, np.integer):
-                assert av.min() >= 0 and av.max() <= AVLabel.UNK, "Invalid vessels map"
-                av = av.astype(np.uint8)
+                    av_ = av_[::-1]  # Invert the channels: red (artery) and blue (vein)
+                av_map[av_[0]] = AVLabel.ART
+                av_map[av_[2]] = AVLabel.VEI
+                av_map[av_[1]] = AVLabel.BOTH
+                av_map[av_[2] & av_[0]] = AVLabel.BOTH
+                av_ = av_map
+        elif av_.ndim == 2:
+            if av_.dtype == bool:
+                av_ = (av_ * AVLabel.UNK).astype(np.uint8)
+            elif np.issubdtype(av_.dtype, np.integer):
+                assert av_.min() >= 0 and av_.max() <= AVLabel.UNK, "Invalid vessels map"
+                av_ = av_.astype(np.uint8)
             else:
                 raise ValueError("The vessels map must be a binary image or a label map using the AVLabel convention.")
 
         # --- Resize the image ---
-        if target_shape is not None and av.shape != target_shape:
+        if target_shape is not None and av_.shape != target_shape:
             match ReshapeMethod.parse(reshape_method):
                 case ReshapeMethod.RESIZE:
-                    vessel_mask = av > 0
+                    vessel_mask = av_ > 0
                     vessel_mask = resize(vessel_mask, target_shape, interpolation=True)
-                    av = resize(av, target_shape, interpolation=False) * vessel_mask
-                    # TODO: Smooth av boundaries (right now, only vessel_mask is smoothed)
+                    av_ = resize(av_, target_shape, interpolation=False) * vessel_mask
+                    # TODO: Smooth av_ boundaries (right now, only vessel_mask is smoothed)
                 case ReshapeMethod.CROP:
-                    av = crop_pad_center(av, target_shape)
+                    av_ = crop_pad_center(av_, target_shape)
                 case ReshapeMethod.RAISE:
                     raise ValueError(
-                        f"The artery/vein maps shape {av.shape} differs from the fundus image shape {target_shape}."
+                        f"The artery/vein maps shape {av_.shape} differs from the fundus image shape {target_shape}."
                     )
 
-        return av  # type: ignore
+        return av_  # type: ignore
 
     @overload
     @classmethod
     def load_od_macula(
         cls,
-        seg,
+        seg: ImageSource,
         target_shape: Optional[Tuple[int, int]] = None,
         *,
-        reshape_method: ReshapeMethod = ReshapeMethod.RAISE,
+        reshape_method: ReshapeMethods = ReshapeMethod.RAISE,
         fit_ellipse: Literal[False] = False,
     ) -> Tuple[npt.NDArray[np.bool_], Point]: ...
     @overload
     @classmethod
     def load_od_macula(
         cls,
-        seg,
+        seg: ImageSource,
         target_shape: Optional[Tuple[int, int]] = None,
         *,
-        reshape_method: ReshapeMethod = ReshapeMethod.RAISE,
+        reshape_method: ReshapeMethods = ReshapeMethod.RAISE,
         fit_ellipse: Literal[True],
     ) -> Tuple[npt.NDArray[np.bool_], Point, Point]: ...
     @classmethod
     def load_od_macula(
         cls,
-        seg,
+        seg: ImageSource,
         target_shape: Optional[Tuple[int, int]] = None,
         *,
-        reshape_method: ReshapeMethod = ReshapeMethod.RAISE,
+        reshape_method: ReshapeMethods = ReshapeMethod.RAISE,
         fit_ellipse: bool = False,
     ) -> Tuple[npt.NDArray[np.bool_], Point] | Tuple[npt.NDArray[np.bool_], Point, Point]:
         # --- Load the image ---
         if isinstance(seg, (str, Path)):
-            seg = read_image(seg, cast_to_float=False)
+            seg_ = read_image(seg, cast_to_float=False)
         elif is_torch_tensor(seg):
-            seg = seg.detach().cpu().numpy()
+            seg_ = seg.numpy(force=True)
+        elif isinstance(seg, np.ndarray):
+            seg_ = seg
+        else:
+            raise TypeError("The optic disc segmentation must be a path, a numpy array or a torch tensor.")
 
         # --- Interpret image ---
-        assert isinstance(seg, np.ndarray), "The optic disc or macula map must be a numpy array."
-        if seg.ndim == 3:
+        if seg_.ndim == 3:
+            if seg_.shape[0] not in (1, 3) and seg_.shape[2] in (1, 3):
+                seg_ = seg_.transpose(2, 0, 1)  # HWC -> CHW
             # Check if the image is a binary image
-            MAX = 255 if seg.dtype == np.uint8 else 1
-            seg = np.any(seg > MAX / 2, axis=2)
-        assert seg.ndim == 2, "The optic disc map must be a grayscale image."
+            MAX = 255 if seg_.dtype == np.uint8 else 1
+            seg_ = np.any(seg_ > MAX / 2, axis=0)
+        assert seg_.ndim == 2, "The optic disc map must be a grayscale image."
 
-        if seg.dtype != bool:
-            MAX = 255 if seg.dtype == np.uint8 else 1
-            seg = seg > MAX / 2
+        if seg_.dtype != bool:
+            MAX = 255 if seg_.dtype == np.uint8 else 1
+            seg_ = seg_ > MAX / 2
 
         # --- Resize the image ---
-        if target_shape is not None and seg.shape != target_shape:
+        if target_shape is not None and seg_.shape != target_shape:
             match ReshapeMethod.parse(reshape_method):
                 case ReshapeMethod.RESIZE:
-                    seg = resize(seg, target_shape, interpolation=True)
+                    seg_ = resize(seg_, target_shape, interpolation=True)
                 case ReshapeMethod.CROP:
-                    seg = crop_pad_center(seg, target_shape)
+                    seg_ = crop_pad_center(seg_, target_shape)
                 case ReshapeMethod.RAISE:
                     raise ValueError(
-                        f"The artery/vein maps shape {seg.shape} differs from the fundus image shape {target_shape}."
+                        f"The artery/vein maps shape {seg_.shape} differs from the fundus image shape {target_shape}."
                     )
 
         # --- Find the centroid ---
-        return find_centroid(seg, fit_ellipse=fit_ellipse)
+        return find_centroid(seg_.astype(np.bool_), fit_ellipse=fit_ellipse)
 
     ####################################################################################################################
     #    === PROPERTY ACCESSORS ===
@@ -515,14 +624,14 @@ class FundusData:
 
     @property
     def has_image(self) -> bool:
-        return self._fundus is not None
+        return self._image is not None
 
     @property
     def image(self) -> npt.NDArray[np.float32]:
         """The fundus image in RGB format. Values are float32 in the range [0, 1]."""
-        if self._fundus is None:
+        if self._image is None:
             raise AttributeError("The fundus image was not provided.")
-        return self._fundus
+        return self._image
 
     @property
     def has_fundus_mask(self) -> bool:
@@ -679,7 +788,11 @@ class FundusData:
         return mask
 
     @property
-    def od_size(self) -> Optional[Point]:
+    def has_od_size(self) -> bool:
+        return self._od_size is not None and self._od_size is not ABSENT
+
+    @property
+    def od_size(self) -> Point:
         """The size (width, height) of the optic disc or None if the optic disc is not visible in this fundus.
 
         Raises
@@ -692,18 +805,24 @@ class FundusData:
                 raise AttributeError("The optic disc segmentation was not provided.")
             else:
                 _, self._od_center, self._od_size = self.load_od_macula(self._od, self.shape, fit_ellipse=True)
-        return None if self._od_size is ABSENT else self._od_size
+        if self._od_size is ABSENT:
+            raise AttributeError("The optic disc is not visible in this fundus.")
+        return self._od_size
 
     @property
-    def od_diameter(self) -> Optional[float]:
-        """The diameter of the optic disc or None if the optic disc is not visible in this fundus.
+    def has_od_diameter(self) -> bool:
+        return self.has_od_size
+
+    @property
+    def od_diameter(self) -> float:
+        """The diameter of the optic disc or -1 if the optic disc is not visible in this fundus.
 
         Raises
         ------
         AttributeError
             If the optic disc segmentation was not provided.
         """
-        return None if self._od_size is None else self._od_size.max
+        return -1 if self._od_size is None else self._od_size.max
 
     @property
     def has_macula(self) -> bool:
@@ -741,7 +860,7 @@ class FundusData:
     def infered_macula_center(self) -> Optional[Point]:
         """The center of the macula or the center of the fundus if the macula is not visible."""
         if self._macula_center is not None:
-            return None if self._macula_center.is_nan() else self._macula_center
+            return None if self._macula_center is ABSENT else self._macula_center
         if self._od_center is None:
             return None
 
@@ -764,9 +883,75 @@ class FundusData:
 
         return self._name
 
+    @property
+    def mutable(self) -> bool:
+        """Whether this FundusData instance is immutable."""
+        return not self._immutable
+
+    @property
+    def immutable(self) -> bool:
+        """Whether this FundusData instance is immutable."""
+        return self._immutable
+
+    ####################################################################################################################
+    #    === EXPORT METHODS ===
+    ####################################################################################################################
+    def write_image(
+        self,
+        image: Optional[PathLike] = None,
+        od: Optional[PathLike] = None,
+        macula: Optional[PathLike] = None,
+        vessels: Optional[PathLike] = None,
+        av: Optional[PathLike] = None,
+        *,
+        overwrite: bool = False,
+    ) -> None:
+        """Save the fundus image and the provided segmentations to files.
+
+        Parameters
+        ----------
+        image : PathLike
+            If provided, save the fundus image at the given path. If the path correspond to a directory, the image is saved in a file named `<name>.jpg` in that directory.
+
+        od : PathLike
+            If provided, save the optic disc segmentation at the given path. If the path correspond to a directory, the segmentation is saved in a file named `<name>.png` in that directory.
+        macula : PathLike
+            If provided, save the macula segmentation at the given path. If the path correspond to a directory, the segmentation is saved in a file named `<name>.png` in that directory.
+        vessels : PathLike
+            If provided, save the vessels segmentation at the given path. If the path correspond to a directory, the segmentation is saved in a file named `<name>.png` in that directory.
+        av : PathLike
+            If provided, save the arteries and veins segmentation at the given path. If the path correspond to a directory, the segmentation is saved in a file named `<name>.png` in that directory.
+
+        overwrite : bool, optional
+            Whether to overwrite the files if they already exist, by default False.
+        Raises
+        ------
+        AttributeError
+            If one of the requested data to save was not provided.
+        """  # noqa: E501
+        on_exists = "overwrite" if overwrite else "ignore"
+        if image is not None:
+            write_image(self.image, image, default_filename=f"{self.name}.jpg", on_exists=on_exists)
+        if od is not None:
+            write_image(self.od, od, default_filename=f"{self.name}.png", on_exists=on_exists)
+        if macula is not None:
+            write_image(self.macula, macula, default_filename=f"{self.name}.png", on_exists=on_exists)
+        if vessels is not None:
+            write_image(self.vessels, vessels, default_filename=f"{self.name}.png", on_exists=on_exists)
+        if av is not None:
+            av_color = label_map_to_rgb(self.av_not_masked, typing.cast(dict[int, np.ndarray], FundusData.AV_COLORS))
+            write_image(av_color, av, default_filename=f"{self.name}.png", on_exists=on_exists)
+
     ####################################################################################################################
     #    === VISUALISATION TOOLS ===
     ####################################################################################################################
+    AV_COLORS = {
+        AVLabel.ART: np.array([255, 0, 0], np.uint8),  # Red
+        AVLabel.VEI: np.array([0, 0, 255], np.uint8),  # Blue
+        AVLabel.BOTH: np.array([255, 0, 255], np.uint8),  # Magenta
+        AVLabel.UNK: np.array([0, 255, 0], np.uint8),  # Green
+    }
+
     def draw(self, labels_opacity=0.5, *, view=None):
         from jppype import Mosaic, View2D
 
@@ -776,7 +961,7 @@ class FundusData:
             return view
         elif view is None:
             view = View2D()
-        view.add_image(self._fundus, "fundus")
+        view.add_image(self._image, "fundus")
 
         COLORS = {
             AVLabel.ART: "coral",
@@ -784,7 +969,7 @@ class FundusData:
             AVLabel.BOTH: "darkorchid",
             AVLabel.UNK: "gray",
             10: "white",
-            11: "white",
+            11: "teal",
         }
 
         if self._av is not None:
@@ -824,8 +1009,8 @@ class FundusData:
 
         other = copy(self)
 
-        if self._fundus is not None:
-            other._fundus = rotate(self._fundus, angle)
+        if self._image is not None:
+            other._image = rotate(self._image, angle)
         if self._fundus_mask is not None:
             other._fundus_mask = rotate(self._fundus_mask, angle)
         if self._vessels is not None:
@@ -863,8 +1048,8 @@ class FundusData:
             raise
 
         other = copy(self)
-        if self._fundus is not None:
-            other._fundus = resize(self._fundus, shape, interpolation=True)
+        if self._image is not None:
+            other._image = resize(self._image, shape, interpolation=True)
         if self._fundus_mask is not None:
             other._fundus_mask = resize(self._fundus_mask, shape, interpolation=True)
         if self._vessels is not None:
